+++ src/main/java/org/service/ExchangeRateService.java
@@ -2,6 +2,8 @@ package org.service;
 
 import org.dto.ExchangeRateDTO;
 import org.model.ExchangeRate;
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,7 +12,7 @@ public interface ExchangeRateService {
     Optional<ExchangeRate> getExchangeRate(String base_currency_code, String target_currency_code);
     List<ExchangeRateDTO> getAllExchangeRates();
     void update(String base_currency_code, String target_currency_code, double rate);
-    void save(String base_currency_code, String target_currency_code, double rate);
+    void save(String base_currency_code, String target_currency_code, BigDecimal rate);
 
 
 }
