+++ src/main/java/org/service/ExchangeService.java
@@ -2,6 +2,8 @@ package org.service;
 
 import org.dto.ExchangeResponseDTO;
 
+import java.math.BigDecimal;
+
 public interface ExchangeService {
-    ExchangeResponseDTO exchange(String baseCurrencyCode, String TargetCurrencyCode, double amount);
+    ExchangeResponseDTO exchange(String baseCurrencyCode, String TargetCurrencyCode, BigDecimal amount);
 }
